List-master
This project is a web application that allows users to upload an Excel file containing data, combine it with user-entered data using JavaScript, and then save the combined data into an SQL database. Additionally, the project provides a functionality to delete the SQL table and convert it back to an Excel file for download.